backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
httpx>=0.25.1
geopy>=2.3.0
anthropic>=0.54.0


